docker/docker_c/parallel/requirements.txt
# requirements.txt
numpy==1.26.4
awkward==2.7.0            
uproot==5.3.0            
pika==1.3.2
matplotlib==3.8.3
python-dotenv==1.0.0
vector==1.6.1
pandas==2.2.1
requests == 2.32.3
aiohttp == 3.9.5
pyarrow==14.0.2              